+++ cmd/gale/audit_fix_U9_test.go
@@ -126,10 +126,8 @@ func TestRemoveHostFlagRemovesForeignHostEntry(t *testing.T) {
 	t.Cleanup(func() { os.Chdir(orig) })
 
 	removeGlobal = true
-	removeHost = "otherbox"
 	t.Cleanup(func() {
 		removeGlobal = false
-		removeHost = ""
 	})
 
 	if err := removeCmd.RunE(removeCmd, []string{"foo"}); !errors.Is(err, errSwitchHosts) {